requirements.txt
# Core runtime
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.8.0


# NLP + intent + dates
spacy==3.7.5
transformers==4.44.2
torch==2.4.0
sentencepiece==0.2.0
huggingface-hub==0.25.2


dateparser==1.2.0


# UI
streamlit==1.38.0
requests==2.32.3


# Optional extras you can enable later
nltk==3.9.1 # VADER sentiment
scikit-learn==1.5.2 # IsolationForest etc
pytesseract==0.3.13 # OCR
pdfplumber==0.11.4 # PDF policy extraction
chromadb==0.5.5 # local vector DB for RAG
langchain==0.2.15